+++ error/options_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/go-kit/kit/log"
+	"github.com/goph/emperror"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -12,4 +13,11 @@ func TestLogger(t *testing.T) {
 	opts := newOptions(Logger(logger))
 
 	assert.Equal(t, logger, opts.logger)
+}
+
+func TestHandler(t *testing.T) {
+	handler := new(emperror.TestHandler)
+	opts := newOptions(Handler(handler))
+
+	assert.Equal(t, handler, opts.handlers[0])
 }