+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
 
+## Unreleased
+
+### Added
+
+- Additional handler options to error handler
+
+
 ## 0.1.0 - 2017-10-01
 
 - Initial release